+++ spec/shelly/cli/organization_spec.rb
@@ -10,7 +10,7 @@ describe Shelly::CLI::Organization do
   Shelly::Client.stub(:new).and_return(@client)
   $stdout.stub(:puts)
   $stdout.stub(:print)
-  @client.stub(:token).and_return("abc")
+  @client.stub(:authorize!)
   FileUtils.mkdir_p("/projects/foo")
   Dir.chdir("/projects/foo")
   end